feat: add support for hexes in constants
